+++ src/compas/data/encoders.py
@@ -40,7 +40,7 @@ except (ImportError, SyntaxError):
     numpy_support = False
 
 
-def cls_from_dtype(dtype, mro=None):  # type: (...) -> Type[Data]
+def cls_from_dtype(dtype, inheritance=None):  # type: (...) -> Type[Data]
     """Get the class object corresponding to a COMPAS data type specification.
 
     Parameters
@@ -48,8 +48,8 @@ def cls_from_dtype(dtype, mro=None):  # type: (...) -> Type[Data]
     dtype : str
         The data type of the COMPAS object in the following format:
         '{}/{}'.format(o.__class__.__module__, o.__class__.__name__).
-    mro : list[str], optional
-        The MRO of the class, all superclasses of the class that can be used if given dtype is not found.
+    inheritance : list[str], optional
+        The inheritance chain of this class, a list of superclasses that can be used if given dtype is not found.
 
     Returns
     -------
@@ -66,12 +66,12 @@ def cls_from_dtype(dtype, mro=None):  # type: (...) -> Type[Data]
 
     """
 
-    if mro is None:
-        full_mro = [dtype]
+    if inheritance is None:
+        full_inheritance = [dtype]
     else:
-        full_mro = [dtype] + mro
+        full_inheritance = [dtype] + inheritance
 
-    for dtype in full_mro:
+    for dtype in full_inheritance:
         mod_name, attr_name = dtype.split("/")
         try:
             module = __import__(mod_name, fromlist=[attr_name])
@@ -81,7 +81,7 @@ def cls_from_dtype(dtype, mro=None):  # type: (...) -> Type[Data]
         except AttributeError:
             continue
 
-    raise ValueError("No class found in MRO: {}".format(mro))
+    raise ValueError("No class found in inheritance chain: {}".format(full_inheritance))
 
 
 class DataEncoder(json.JSONEncoder):
@@ -236,7 +236,7 @@ class DataDecoder(json.JSONDecoder):
             return o
 
         try:
-            cls = cls_from_dtype(o["dtype"], o.get("mro", None))
+            cls = cls_from_dtype(o["dtype"], o.get("inheritance", None))
 
         except ValueError:
             raise DecoderError(