+++ src/compas/datastructures/datastructure.py
@@ -21,15 +21,25 @@ class Datastructure(Data):
         self._aabb = None
         self._obb = None
 
-    def __get_mro__(self):
-        mro = []
+    @property
+    def __inheritance__(self):
+        """Get the inheritance chain of the datastructure.
+        Until one level above the Datastructure class (eg. Mesh, Graph, ...).
+
+        Returns
+        -------
+        list[str]
+            The inheritance chain of the datastructure.
+
+        """
+        inheritance = []
         for cls in self.__class__.__mro__:
             if cls == self.__class__:
                 continue
             if cls == Datastructure:
                 break
-            mro.append(cls.__cls_dtype__())
-        return mro
+            inheritance.append(cls.__clstype__())
+        return inheritance
 
     def __jsondump__(self, minimal=False):
         """Return the required information for serialization with the COMPAS JSON serializer.
@@ -47,7 +57,7 @@ class Datastructure(Data):
         state = {
             "dtype": self.__dtype__,
             "data": self.__data__,
-            "mro": self.__get_mro__(),
+            "inheritance": self.__inheritance__,
         }
         if minimal:
             return state